Move docs from GitHub Pages to PyPI :boom:
After finally getting automated docs publishing to GitHub Pages working
on Travis CI, we realized the PyPI deployment provider automatically
uploads docs via PyPI to pythonhosted.org (which we can't seem to
disable).

Which is just as easy, so we can now do away with the GitHub deploy key
and the gh-pages branch.